Removing model namespace
