+++ lib/keycloak_oauth/configuration.rb
@@ -1,0 +1,9 @@
+require 'singleton'
+
+module KeycloakOauth
+  class Configuration
+    include Singleton
+
+    attr_accessor :auth_url, :realm, :client_id, :client_secret
+  end
+end
